Fix delete_events (it was failing because of invalid syntax, "min_stamp, max_stamp = <SQLite row>"). Also ensure that it doesn't explode if it's called with invalid IDs.

+++ _zeitgeist/engine/main.py
@@ -438,10 +438,11 @@ class ZeitgeistEngine:
 			FROM event
 			WHERE id IN (%s)
 		""" % ",".join(["?"] * len(ids)), ids)
-		min_stamp, max_stamp = self._cursor.fetchone()
-	
-		# FIXME: Delete unused interpretation/manifestation/text/etc.
-		self._cursor.execute("DELETE FROM event WHERE id IN (%s)"
-			% ",".join(["?"] * len(ids)), ids)
+		timestamps = self._cursor.fetchone()
 		
-		return min_stamp, max_stamp
+		if timestamps:
+			# FIXME: Delete unused interpretation/manifestation/text/etc.
+			self._cursor.execute("DELETE FROM event WHERE id IN (%s)"
+				% ",".join(["?"] * len(ids)), ids)
+		
+		return timestamps
